Expose two private methods as public
This is to fix that the vendored copy of Content Publisher is exposing
a private method (insertText) and had deviated from the version supplied
in this repo.

This adds this method exposure into the source code and adds another
one: newlinesToSurroundSelectedText, which can be used to determine the
spacing around a block element.

It isn't ideal to be doing this as this only deviates this fork from
upstream. Hopefully in the future Content Publisher has different means
to insert text that don't involve using this private API.