Add testing for osx keydown browser bug
On macOS, there is an existing bug that will always
report the keydown event.key as lowercase, even
when pressing cmd-shift-letter.

This can result in the toolbar hotkey firing for
both the uppercase and lowercase meta keys.  This
was fixed in a prior commit, but a test is being
added here to emulate this behavior in the test,
so that there is a “breaking” regression test.