Copy dist files into examples directory
This allows the directory examples to be rendered without requiring
contents from the rest of this repo. This change allows GitHub pages to
work properly as they won't support a relative link down a directory.

This directory is ignored from git since it is generated. Ideally the
other dist directory in this repo would be git ignored too, but
resolving this is beyond this commit's scope.